+++ examples/inner-tx-create-assets/scripts/utils/index.js
@@ -1,0 +1,14 @@
+const accounts = (deployer) => {
+	return {
+		creator: deployer.accountsByName.get("master-account"),
+	};
+};
+
+const decodeValue = (value) => {
+	return new TextDecoder().decode(value);
+};
+
+module.exports = {
+	accounts,
+	decodeValue,
+};