+++ packages/algob/src/lib/tx.ts
@@ -267,54 +267,56 @@ export function signTransactions(txnAndSign: wtypes.TransactionAndSign[]): Uint8
 	});
 }
 
-// /**
-//  * Execute single transaction or group of transactions (atomic transaction)
-//  * executes `ExecParams` or `Transaction` Object, SDK Transaction object passed to this function
-//  * will be signed and sent to network. User can use SDK functions to create transactions.
-//  * Note: If passing transaction object a signer/s must be provided.
-//  * @param deployer Deployer
-//  * @param transactionParam transaction parameters or atomic transaction parameters
-//  * https://github.com/scale-it/algo-builder/blob/docs/docs/guide/execute-transaction.md
-//  * or TransactionAndSign object(SDK transaction object and signer parameters)
-//  */
-// export async function executeTx(
-// 	deployer: Deployer,
-// 	transactions: wtypes.ExecParams[] | wtypes.TransactionAndSign[]
-// ): Promise<ConfirmedTxInfo> {
-// 	let isSDK = false;
-// 	let signedTxn;
-// 	if (transactions.length === 0) {
-// 		throw new BuilderError(ERRORS.GENERAL.EXECPARAMS_LENGTH_ERROR);
-// 	}
-// 	if (wtypes.isSDKTransactionAndSign(transactions[0])) {
-// 		signedTxn = signTransactions(transactions as wtypes.TransactionAndSign[]);
-// 		isSDK = true;
-// 	}
+/**
+ * Execute single transaction or group of transactions (atomic transaction)
+ * executes `ExecParams` or `Transaction` Object, SDK Transaction object passed to this function
+ * will be signed and sent to network. User can use SDK functions to create transactions.
+ * Note: If passing transaction object a signer/s must be provided.
+ * @param deployer Deployer
+ * @param transactionParam transaction parameters or atomic transaction parameters
+ * https://github.com/scale-it/algo-builder/blob/docs/docs/guide/execute-transaction.md
+ * or TransactionAndSign object(SDK transaction object and signer parameters)
+ */
+export async function executeTx(
+	deployer: Deployer,
+	transactions: wtypes.ExecParams[] | wtypes.TransactionAndSign[]
+): Promise<ConfirmedTxInfo[]> {
+	let isSDK = false;
+	let signedTxn;
+	if (transactions.length === 0) {
+		throw new BuilderError(ERRORS.GENERAL.EXECPARAMS_LENGTH_ERROR);
+	}
+	if (wtypes.isSDKTransactionAndSign(transactions[0])) {
+		signedTxn = signTransactions(transactions as wtypes.TransactionAndSign[]);
+		isSDK = true;
+	}
 
-// 	if (isSDK && signedTxn) {
-// 		const confirmedTx = await deployer.sendAndWait(signedTxn);
-// 		console.debug(confirmedTx);
-// 		return confirmedTx;
-// 	}
+	if (isSDK && signedTxn) {
+		await deployer.sendAndWait(signedTxn);
+		return await deployer.getReceiptTxns(
+			(transactions as wtypes.TransactionAndSign[]).map((txn) => txn.transaction)
+		);
+	}
 
 // 	const execParams = transactions as wtypes.ExecParams[];
 
-// 	deployer.assertCPNotDeleted(execParams);
-// 	try {
-// 		const txIdxMap = new Map<number, [string, wtypes.ASADef]>();
-// 		const [txns, signedTxn] = await makeAndSignTx(deployer, execParams, txIdxMap);
-// 		const confirmedTx = await deployer.sendAndWait(signedTxn);
-// 		console.debug(confirmedTx);
-// 		if (deployer.isDeployMode) {
-// 			await registerCheckpoints(deployer, execParams, txns, txIdxMap);
-// 		}
-// 		(console.log as any).restore();
-// 		console.log(confirmedTx);
-// 		return confirmedTx;
-// 	} catch (error) {
-// 		if (deployer.isDeployMode) {
-// 			deployer.persistCP();
-// 		}
+	deployer.assertCPNotDeleted(execParams);
+	try {
+		const txIdxMap = new Map<number, [string, wtypes.ASADef]>();
+		const [txns, signedTxn] = await makeAndSignTx(deployer, execParams, txIdxMap);
+		await deployer.sendAndWait(signedTxn);
+		const confirmedTx = await deployer.getReceiptTxns(txns);
+		console.debug(confirmedTx);
+		if (deployer.isDeployMode) {
+			await registerCheckpoints(deployer, execParams, txns, txIdxMap);
+		} else {
+			console.warn("deploy app/asset not allowed in run mode");
+		}
+		return confirmedTx;
+	} catch (error) {
+		if (deployer.isDeployMode) {
+			deployer.persistCP();
+		}
 
 // 		throw error;
 // 	}